Feature: pasting hardlinked trees
This feature is useful when there is a need to create a directory tree,
which is similar to an existing one while keeping all the files in this
new tree hardlinked to those existing ones. Such functionality might be
useful when rearranging a huge directory tree (or a set of directory
trees) into some new layout.

This operation is done synchronously (i.e. it doesn't use CommandLoader
to put the ongoing operation into the task queue).

Target end of the operation is handled the following way:

1. If target directory with the same name exists - it gets used.

2. If target folder does not exist - it is created with the same mode
   bits as corresponding source one.

3. If target file with the same name exists and is a hardlink of the
   source file already - it is getting left as it is. If it's not a
   hardlink of the source file - a new hardlink is created in a form of
   <original>_, <original>_1, ..., etc.

4. If target file does not exist - a hardlink is created.