ext.rifle: flags now work with commands with pipes, semicolons etc
rough explanation: before, the "t" flag would run the program in a
terminal by prepending something like "xterm -e " to the command.
If the command is "ls | less", it would result in "xterm -e ls | less".
This commit changes it so the result looks more like
"xterm -e sh -c 'ls | less'" and works as intended.