core.actions: imported shutil_g paste function from v1.2.3
The code implements the functionality of cp and mv coreutils in python
and was removed as of v1.4.0 for efficiency and simplicity reasons.
I moved it back in for several reasons:

1. I plan to enhance shutil_g to report on its copying status so I can
display a progress bar

2. With no need for external cp/mv programs, distribution-specific
differences (like broken backup option on MacOS/BSD?) become irrelevant

3. You can still copy in ranger if you accidently delete /bin/cp

4. It gets rid of the occasional bug that the cp process keeps running
when ranger is terminated while copying

The possible downside is reduced efficiency in copying and ranger might
get stuck if the copying function is blocked.  Let's see if it works
out.