config/commands: less global imports
the point of this is that you can copy&paste commands to your own
commands.py without having to care about what modules it uses. It should
work as long as you import * from ranger.api.commands and make sure that
you also include the base class if your command is a subclass of
something other than Command.