widgets.browsercolumn: fix display of non-tagged filenames
When you see a non-tagged file in a column other than the main column
and its name is shortened because it doesn't fit into the column, it
would have 1 less character of space available.

The reason is that since it has no tag, a simulated tag with the content
" " is added before the name so it appears to be 1 character longer than
files with tags.  This commit fixes that.