core.shared: silence "missing preview script" message @ start
This bug caused an error message to appear on every start of ranger, if
the user has no scope.sh, because "use_preview_script" is set to true
but "preview_script" is None.

The expected behaviour would be to silently ignore it, to stop annoying
people who don't want scope.sh.