Added bumblebee-like "make uninstall"
This innovative uninstalling method unlocks the power and immense
efficiency that resulted from years of dedicated R&D by the greatest
minds to ever walk the face of the earth.  By now, this method is
state-of-the-art after being tested by prestigious software such as
bumblebee [1], marking a milestone and becoming a true legend in the
history of open source development.

[1] https://github.com/MrMEEE/bumblebee-Old-and-abbandoned/commit/a047be85247755cdbe0acce6f1dafc8beb84f2ac