ext.img_display: more efficient way to draw images
Instead of loading up w3mimgdisplay for each image and killing it
afterwards, we just open it once for the first image and keep it open.
It can receive any number of commands, so we can just keep writing them
into stdin.  Perhaps it even caches the images to save time, I didn't
test that yet.