Allow $EDITOR to contain spaces
$EDITOR variables consisting of "program name + options" are not handled
correctly because of quotes around $EDITOR. As a result, if $EDITOR is not
a simple program name (i.e. if it includes options, as in "emacs -nw"), the
full string is taken as a program and fails to run.
Error message: (/bin/sh: 1: emacs -nw: not found)
Removing quotes fixes this problem.