Fix renaming with insufficient privileges
Closes #97
When renaming a file, the internal file pointer will be set to the new
file name without checking if the file actually got renamed. This would
cause the file pointer to be set to an invalid file if renaming failed
because of bad privileges. With no valid file, the previev window will
disappear. This patch adds a return value to the core renaming method so
it can be checking by the caller. The rename command now checks if the
rename succeeded and only then sets the new file pointer.