core.fm: more sane --copy-config of commands.py
This commit makes ranger only copy a small sample commands.py to
~/.config/ranger/commands.py when running --copy-config=commands.
The default commands.py will be copied to commands_full.py instead, a
file that will not be loaded at all, and will only serve as a reference.
It can be obtained with --copy-config=commands_full

This will avoid that the average user's commands.py will break ranger
whenever she's updating ranger after using --copy-config=all.
It's necessary because commands.py contains much of the actual
program logic, but it can't be kept compatible like other configuration
files, and the user can't be expected to always keep an up-to-date
version of it.

Overview:

Removed --copy-config rule:

--copy-config=commands:  commands.py -> commands.py

New --copy-config rules:

--copy-config=commands:      commands_sample.py -> commands.py
--copy-config=commands_full: commands.py -> commands_full.py