applied generally useful changes from emacs branch
The 'lt' and 'gt' special keys were added so you can bind actions to
something like "<C-lt>" meaning "Ctrl+<" and <C-gt> meaning "Ctrl+>".

"key > 0" was changed to "key >= 0" in ranger.gui.ui so that Ctrl+Space
(which has the key value 0) can be recognized.  I am not 100% sure if
this has side effects.