Use setuptools if present
The main reason for this is the presence of `develop` command in setuptools.

So by running `sudo python setup.py develop` - ranger gets installed, but you
can easily make changes to the code and won't have to reinstall it.

You can read more about it here:
http://pythonhosted.org//setuptools/setuptools.html#development-mode

It basically links an egg, so I had to add an appropriate line to .gitignore
as well.