Fix for issue #332. Broken file shortening.
The filename on the second to last line of the screen should now be
properly shortened. It would seem the function addstr can fail after
having already written some characters and therefore moved the
cursor. We now keep track of the cursor's starting position so that
in case of such an exception, we can move the cursor back to where it
started.