ext/rifle.py: Fix the "X" specifier for Mac OS X
On Mac OS X the "$DISPLAY" environmental variable is never set. Checking
for it to check if there is a graphical environment available will not
work. I think it can be safely assumed that Mac OS X is always running
with GUI.

Calling it "X" is a bit of a misnomer considering Mac OS X does not use
X11, but I think it still conveys the meaning well.