Refactor and improve the TERMCMD handling
Now only the first token of TERMCMD needs to exist in PATH. Previously
the whole value was checked, which could contain additional parameters.

The check and term guessing were performed in two places. Now it's done
by the get_term() function to reduce the code duplication. In that
function the second check was indented one level deeper (relative to the
original code). Keeping both checks on the same level, while visually
pleasing, was conveying the wrong message. One never wants to perform
the second check if the first one fails. Nesting shouldn't be a problem:
if the nesting will need to be any deeper than this, this code will need
to be refactored anyway.

Fixes #613.