Fix cursor being lost when changing viewmodes
This results from finalize being called in order for the widgets.
Previously when ranger changed viewmodes, it would delete then readd the widget
object which meant that the finalize for the viewmode was called last instead
of before console, which meant the finalize of the viewmode was setting
and overriding the cursor position.