Do not suppress custom commands import exceptions
Original problem
----------------
All the import exceptions raised while trying to load custom commands
file were suppressed. This make difficult to understand why the custom commands
file was not loaded.

Fix
---
In the case of import exception , an explanatory message is printed
on stderr and the original exception is raised.

Bonus
-----
The path of custom commands file that caused the exception is printed
in the explanatory message.