container.fsobject: Fix some issues with symlink stat handling
In the file loading code, if the loaded file was a symlink, the last
stored modification time of the real file (`self.stat`) was compared
with the modification time of the symlink (`real_ctime`). It was
leading to some subtle issues. The most visible result of this bug was
a flickering image preview of the symlinked images.

The bug was introduced in ec5dfcb1cb0f549fea3ecf7b0505f02e197887b2.