ext.lazy_property: don't redefine resetter function
Without this change, binding a signal to a resetter function with a weak
reference would delete the signal after it's called, because the
function would be disconnected from the object and market for garbage
collection.

Also adapted the test cases to check for this problem.

For example, commit 22ffd78ef654cd7ae4357a908408ac19 binds vcs__reset()
to the signal "setopt.vcs_aware" with a weak ref.  Because of that, the
signal is only triggered the first time the setting is changed.