scope.sh: Add the image-based PDF previews (disabled by default)
By default they are rendered for the FullHD resolution as it should be
both legible on any resolution and generate not too large files.

The order of handing the previews was changed to handle the
image-based previews first as they are easy to toggle anyway and it
makes sense for them to override any other preview type if enabled and
available.