Prefer VISUAL to EDITOR
`VISUAL` is often misused to mean desktop GUI editor it's actual purpose
was to distinguish editors like `vi` (which is short for visual) from
editors like `ed` (which was probably short for editor), `EDITOR` is
actually intended to point to one of the latter.

Rifle still defaults to `vim` if both variables are undefined, even
with an old `rifle.conf` after trying to fall back on `EDITOR`.

For some better information on this see,
https://unix.stackexchange.com/a/334022