Guard tmux title changes
Changing the tmux window title to "ranger" was not guarded for the
presence of the tmux executable, this tripped up at least one user with
an `rc.conf` that still `set update_tmux_title true`.
While the behavior is *not-a-bug*, I expect most people'd rather have
the setting enabled by default since "python" is a less useful window
title.

Fix #1042