Update bindings for media directories
Because of `udisks2` `/run/media/$USER` is becoming a common directory
for automounting, this adds a keybinding to go there (using the first
available letter in the path).

Similarly the binding for `/media` now points to `/media/$USER`, because
for example ubuntu adopted this new location.
I believe the `$USER` parts have to do with security, giving only the
user that mounted a filesystem access to it.