Fix scope.sh crashing on text ending with nullbyte
Currently, scope.sh doesn't differentiate between charsets.
So, if a file with the mimetype `text/plain` is encoded in
binary, and contains a trailing nullbyte within the first 40
bytes, it will crash with `embedded null charcater`. This is
due to using `file --dereference --brief --mime-type --` on
line 170, instead of using the POSIX-compatible `file -i`,
which does the same thing, but also lists the charset.

The main culprit is the `file` utility, which lists any files
with 40 or less bytes, even if they contain control characters,
as `ISO-8859 text` or `ASCII text`. If the last byte is a
nullbyte, it will be sent to Python and crash the draw()
function.

If the `MIMETYPE` variable was defined with the second version,
all you would need to do is add a `*binary)` case to
`handle_mime`, right before `text/* | */xml)`, and just let it
fall through to the regular `fallback_handler`. Instead, the
most inexpensive thing you can do is retrieve the first 40 bytes
of any file with mimetype `text/plain` with `head -n 3`. Then
do a hex dump in named-character format, `od -t a`, on the output
and pipe it to `grep -q 'nul'` which will determine if it
contains any nullbytes. If it does, call `return` and exit the
case statement to the `fallback_handler`.

This is also is a very minor security concern.

Resolves: #1094