ranger <path> can select files
Because `ranger somefile` used to invoke rifle but that moved to its own
command `ranger somefile` seemed defunct. People expected ranger to
select the file and were baffled when it crashed. Now the documentation
using the verbiage "path" is no longer confusing/a lie because it just
works™ (until it breaks).

Fixes #930