Revert "ranger <path> can select files"
Due to this commit, pressing `<Right>` on a file does nothing.  Normally
it should open the file.  Until this is fixed, I'll remove this commit
to restore this basic functionality.  Unfortunately, the revert makes
ranger crash when run like `ranger <somefilename>`, but this is a much
less frequent use case, so the revert is still warranted.

This reverts commit 2288a40b45ccdc57d287648c96c5bbe378b95a6d.