In terminals devoid of color cut items were invisible
Since most people use dark background terminals - yes, I'm assuming
because it's my preference - the black color for cut items is nearly
invisible the intention was for bold to be interpreted as bright but
that's a dirty dirty trick. Since dim white sounds like bright black
that's what we fall back to now for terminals that don't support BRIGHT
colors.

The solarized theme remains unchanged, so still uses bold, because it
looks like it knows what it's doing.

Fixes #1185