Use dictionaries for colorscheme definition
Have it use dictionaries.

Because of how colorschemes used to be written, using dictionaries is
backwards compatible because the use function is overwritten. All of the
parsing is done in the use function.

Use dictionaries and tuples to define colorschemes

- Much easier to read than a bunch of if statements
- Still is backwards-compatible with old colorschemes because they
  redefine a function

Edit the colorscheme doc to reflect changes

Fix linting errors and convert jungle colorscheme