Check PYTHON is properly defined
We don't need to check for all the various versions, that's just a
maintenance burden because `python33` shouldn't ever be available if
`python3` isn't.

Added a check for whether `PYTHON` is empty or just whitespace because
that's useless and could actually lead to problems, if `setup.py`
happens to be on the `PATH`.