Add the %any_path macro
The `%any` macro is very useful for commands that know about bookmarks.
Bookmarks are just short names for paths really, so it'd be nice if they
could be used with any commands that take paths as arguments.

Now they can! `%any_pathX` where `X` refers to the which `<any>` the
macro corresponds to (`%any_path` is synonymous with `%any_path0`) is
replaced with the path of the bookmark entered if that bookmark exists.

Fixes #1277