Restore the use of the bold attribute
Brightness by itself is not always clear so in terminals that actually
support boldface it can help to differentiate.

Bright black can be hard to differentiate from white, this is the case
in kitty's default colorscheme. It's still better than the bold black
because while that was easily differentiated from white it was very hard
to see.