Revert "Merge branch 'slackhead-venv-gitignore'"
This reverts commit f5639cd2c4b2f635ec98330da83edbe774a0849d, reversing
changes made to aead73e9488be599e92418cea5d9523da459fde1.

`.venv` is not a canonical name for python virtualenvs so there's no
reason to assume it's common among developers. To avoid potential
surprises because of the ignore we're reverting it.

We recommend using `.git/info/exclude` instead to specify
project-specific ignores that apply only to your setup.