Revert "core.main: reactivate file opening with ranger directly"
Hut misunderstood an issue people were having with ranger's "open
focused on a file" feature and undeprecated the rifle-like
functionality.

The feature's still broken but I'm already redeprecating because
people'll get upset if they start using ranger as rifle and it stops
working again when I fix the feature in the near future.

This reverts commit 67e4ea4e143ed3516e5fe02607bbaf8aedf3534d.