Fix permission confusion when listing dir contents
We were checking the wrong permission for listing directories, runnable
is only the right permission if you want to list the contents of
subdirectories.

We weren't ever updating the permissions but they can change after
loading, that's now improved but not fixed.

Turns out python 2 uses the same OSError for permission issues unless
there's I/O involved as with `open`.

Fixes #1539