Move helper into ranger/ext
There's already a `human_readable` for size in `ranger/ext`, I figure we
might as well add the new helper there, make it easily accessible in
other code.

Did some renaming, no need to drop the `readable` part to shorten the
names. Similarly I changed one of the mappings to be more intuitive.