Abort load if directory is not readable
If a directory turns out not to be readable/`accessible` we need to
abort loading it incrementally.

I tore out `runnable` altogether. It was only erroneously used to check
whether we were able to load a directory. Listing a directory requires
read permission though, not execute permission. `runnable` was relevant
to flattening because you can't list subdirs of a non executable dir but
this caused an infinite recursion. More trouble than it's worth so into
the trash.

Huge thanks to @nQ8hzkZw for pointing out these issues, it takes
dedication to find and reproduce them.

Fixes #1539