Restore pointed_obj when content is loaded
Clearing the `pointed_obj` is somewhat unfortunate because we need to
restore it in some cases. This is currently done by abusing the fact
that `thistab.pathway` isn't cleared. It happens in `load_content` so
the `pointed_obj` doesn't get restored for inaccessible directories.
This is important so the `titlebar` doesn't indicate the wrong path.