Hotfix for an off-by-one in rifle
After #1341 the action IDs displayed by the UI were inconsistent with
the actual numbers used by rifle.  It was caused by the `ask` action
being skipped entirely (including incrementing the ID) but only in the
UI.

This is a hotfix that makes the hidden `ask` action to still take up a
number (usually (always?) 0) and thus keeping the numbering
consistent.  Bad side effect: sometimes the counting starts with 1
instead of 0 but for a time being let's choose a small inconsistency
over a BIG one.