Improve filter_stack documentation
Add an entry about `filter_stack` to the man page, documenting the various
subcommands and available filters. Add documentation about the new
mappings.

Add functionality to the `rotate` subcommand so you can pass the
argument as a `quantifier`, i.e., a numeric prefix. Change the `.r`
mapping to rely on this behavior accordingly, reducing the amount of
unnecessary return pressing.