Refactor to use more idiomatic python
`' '.join(list)` is a very useful and common idiom to join elements of a
list without adding an extra space. In this case `self.rest(1)` gives us
pretty much the same result though.

The `_chaining` recursion was equivalent to a loop over the arguments.

The `if` branches interpreted no arguments as `chain`.