Abstract Loader percent management to ProgressTracker
This makes implementing additional task progress statistics possible. The Loader
doesn't necessarily have to know how to manage the progress percentage
or any other statistic the UI might request. This logic should be
separated.

The bytes_per_tick could not have been used in all places because on
small files (< 16kB), the tick is equal to the file size not to the
bytes_per_tick. This was not a problem when copying one large file
because all the ticks (but the last one) were equal to the
bytes_per_tick. This also makes the variable unused so it was deleted.