Split progress tracking into subclasses
This allows for different status formatting of directory listing and file
copy/move progresses. The directory listing (which uses files as a unit)
should not use the `human_readable()` formatting function. However, the
file copy/move progress (which is in bytes) should use that. Therefore this
could not be solved by simply using different units while having the
same status formatting functions.