trash: Don't call the File() constructor
Previously the File() constructor was called for every path (if the
paths to be moved to trash were supplied after the command instead of
deleting the selection, e.g. ":trash a b c").
This commit adds a method paths_to_filesystem_objects() to find the
existing objects that ranger has in memory and use those instead.