trash: Fix crash on OSError
The trash command used to crash ranger when passing so may arguments
that the argument length limit of the OS is reached.  See the discussion
in pull request #1871 for steps to reproduce.

Now it displays an error message instead of crashing.
(It does not move the files to trash though.)