Improve variable naming in human_readable()
The variable name "byte" doesn't express the contents of the
variable sufficiently. For example "bytes" would be better
but still, these are not bytes (8 bits of data). A plain
number of things is what it is.

This could also be named "file_size" but human_readable
shouldn't know it's being used for file size. In the future
it might be used for something else (e.g. number of bytes transfered).