Implement passing of flags for :bulkrename
The flags a bulkrename command passed to the command runner were fixed
to `w`, which means "wait for an enter keypress." Now it's possible to
pass whichever flags you want. Passing any flag overrides the default so
if you still want the confirmation you need to explicitly pass `w`.

Fixes #2002