Include mtime in hash rather than comparing
Comparing the mtimes of a file and a cached preview or including the
mtime in the preview's hash are almost equivalent except this way we
don't need to find out the preview's mtime.

We include the device identifier and the inode in the hash, this makes
sure duplicate filenames don't cause erroneous hits in the cache.