Drop file path from cache hash
Because we include the device identifier and the inode in cache hashes
now we no longer need the path to the file. Files are already more
uniquely identified (if the path was reused for another file with an
acceptable mtime there used to be a collision).

A small benefit is that hardlinked and symlinked files should now only
require a single cached preview.