Switch from SHA-1 to SHA-512 for the cache
The performance hit is negligible. File name lengths are usually 255
bytes at minimum (except for FAT, but a SHA-1 digest is truncated too).
An attack on the ranger cache based on hash collisions is probably
fairly unlikely but preventing it is almost free.

This should be merged soon after #1838 because both these PRs change the
cache hashes and can therefore cause generation of previews that have
been cached before, wasting some disk space.