Poll rather than use non-blocking I/O
The blocking read doesn't actually pose a problem *if* there is an error
to read. It only gums up the works if there's no error which should mean
the ueberzug process is still running. By polling the process we can see
whether it has stopped running and then read the error.