Switch to packing fields for sha512_encode
The easy solution was to render the `st_dev, st_ino, st_mtime` to a
string and encode that before hashing.

This solution is more efficient because it packs the numbers as bytes
rather than the digits. Technically POSIX.1 does not specify an upper
limit on the sizes for these fields. I think it's safe to say none of
the major operating systems will use more than 8 bytes/64 bits any time
soon.