Fix unicode decode issues in py2
Python 2 implicitly decodes strings as ASCII or the system's default
encoding before encoding a string to bytes. This results in problems if
the string already contains characters outside the ASCII range (> 127).
I reintroduced this bug thinking encoding a string in both python 2 and
3 would be harmless.

Actually swapped the order of the inode and the path which should've
happened in the previous commit.

Fixes #2054