Add bulk command
This is similar to the bulkrename command but it isn't limited to
renaming files, it can change any attribute.

Custom attributes can be added by adding a new class with two methods,
`get_attribute` and `get_change_attribute_cmd`.
The first one receives the file and should return a string representing
the current attribute value.
The second one receives the file, the old attribute value and the new
attribute value as changed by the user, and should return the string
with the command that will be run by the shell script to change the
attribute value from the old one to new one.

Finally, the object should be added as an entry in the `bulk`
dictionary, mapping the attribute name, as it should be passed as a
parameter to the `bulk` ranger command, to the object containing those
two functions.