Allow $PAGER to contain spaces (handle pager flags)
This change is similar to: c826091a74e88102ca1795ab9798dd8b8498f653 .
See the referenced commit for the explanation.

Note that $PAGER is not a strictly better than "$PAGER". In the
cases where the pager executable in PAGER contains spaces e.g.
"/path with spaces/bin/less" the "$PAGER" works correctly and $PAGER
doesn't. This occurrence is much less frequent than pager in PAGER
containing flags, e.g. "less -R" which works correctly with $PAGER, but
doesn't with "$PAGER".

This change is a desirable trade-off.