New feature to print first line of shell command
Now possible to use check_output in order to see the first line printed
to stdout by a given shell command. This is useful when for example
wanting to see the size of a partition without wanting to leave the main
ranger UI.
Example: :check_output du -hs %p
will print the size of the current folder.