Report errors if tag file does not exist
A non-existent tag file leads to an opaque crash without indication to
the user of what went wrong as in #2200. These errors are now reported
through ranger's notify mechanism.

The tag file is no longer created conditionally on `__init__`, only upon
saving tags. This changes the behavior somewhat in that an empty
"tagged" file should never be created.