transpose_words: Fix transposing last two words
Transpose last two words even if on the first letter of last word. This
behavior is consistent with readline and emacs.

The previous code move one word back if it was anywhere in the last word
but that meant the third and second to last words were transposed if the
cursor was on the first character of the last word.