doc: New -s flag for console
Document the new console flag. `-s` allows specifying a sentinel string.
This sentinel is removed from the body of the command and the cursor is
set to its position when opening the console.

Sentinel replacement happens after macro expansion and there's no
characters which cannot occur in a path. It is therefore impossible to
use a sentinel which is 100% "macro safe."

Example, running `console -s <cursor> shell <cursor> %s` will open the
console with the contents `:shell | %s`, where `|` symbolizes the
cursor.