rifle: use default handler for SIGTSTP while running a command (fixes Ctrl-Z bug)
The problem can be fixed (at least for using `vim`) by editing
`ranger/core/ext/rifle.py`. I inserted a little patch to ignore
`SIGTSTP` while the child process is running. Unsure of why this fixes
it, I searched for people experiencing a similar bug and found [this
StackOverflow answer](https://stackoverflow.com/a/31441709) from 2015 (!).

> Saving and restoring the signal handler state as suggested by @lc2817
will work—but it is a workaround because it is not elegant. If
curses.wrapper were modified to add some state to it, to remember if
it was called before, and to save/restore the signal-handlers, the
workaround would be unnecessary. To make it really portable, initscr
should be called on the first use, and refresh on subsequent uses.

Apparently it's an over 5-year old bug in Python's curses library that
hasn't been fixed, which is why this workaround is needed. I'll submit
this workaround as a pull request for now, and see if they will fix this
problem upstream or not.