Remove xargs execution
The use of `xargs` has a limitation on the *length of `argv`* (number of
arguments). After reaching the maximum number of arguments (~850 for me)
`xargs` will still create the new child process, as with the `-n` switch.
`sxiv` (at least for now, `sxiv-26`) is able to read filenames from STDIN on
its own.

This commit by abandoning `xargs` fixes the aforementioned sideeffect, removes
extra dependency and speeds up the work (at least for me).