shell_automatic_cd: Zsh compat
The bind invocation was specific to bash while the function is now POSIX
sh compatible. Since not everyone may want to bind the function to ^O
anyway it is now commented out. With a corresponding comment of how to
do it in zsh. For plain sh one would need to customize `.inputrc`.