browsercolumn: Fix regression when clicking preview
When clicking the preview column ranger used to open the targeted file.
This behavior was broken when implementing preview scrolling.

This reenables that behavior when using the right mouse button. I
decided to deviate from the original behavior because it's prone to
accidentally opening a file when clicking on the terminal window to
bring it to the foreground. This should be an acceptable middle ground.

Fixes #2330