Calculate width of zero-width characters as zero
Unicode is fun and exciting.  Previously, ranger calculated columnar width of the following string as sixteen:

楽しかった♥︎.txt

It now correctly calculates it as fifteen.  Hint:  Open this document in a hex editor or use vim's g8 command to see what hides behind the heart...

Hidden, non-spacing characters seem to be especially popular with girls from Japan.