ui: Fix crash when TMUX not in environment
Made a mistake when merging #2201. If the "TMUX" key isn't in the
environment dictionary we'll get a KeyError. Providing an empty string
default should fix the problem.

I applied the same logic to the check for screen because TERM isn't
guaranteed to be in the environment either, though this verges os
malicious.