commands: Rename setlocal to setinpath
The `setlocal` name is hard to remember, `setintag` is much easier to
recall, so we will follow the same naming style. `setlocal` remains
available as an alias for `setinpath`.

A new abstract class, `_setlocal`, is introduced as a basis for the
`setinpath` command because we want to reuse it for the original
behavior of `setlocal`. Several bug fixes have been integrated in this
new base class. The shifting of arguments was only incidentally correct
most of the time. The matching of quoted arguments only worked in the
absence of nested quotes.