commands: Implement setinregex
The original behavior of `setlocal` allowed to pass in a regular
expression that would be matched to paths. This was confusing for some
users because many paths contain regular expression metacharacters and
these would affect in which directories the settings apply in unexpected
ways. To remedy this we decided to escape all metacharacters by default
but this came at the cost of being unable to specify precisely which
directories would match, as this made use of the `$` metacharacter with
the original behavior.

The behavior has been fixed to never match subdirectories but changing
back to the original behavior and introducing a second backwards
incompatible change was deemed a bad idea. So we decided to make
`setlocal` an alias for the new `setinpath` and implement a `setinregex`
that exhibits the original `setlocal` behavior, allowing for complex
regular expressions.