rifle_sxiv: properly fallback on non-valid extensions
originally we checked the extension to make sure the argument is an
image to avoid edge cases like this:

	./a        # is a dir
	./a.jpg

where running `rifle_sxiv a` would open up `a.jpg` instead of `a` the
directory.

however the is_img function was not correct as it would fail on upper
case extensions, such as `a.JPG`.

instead of doing extension check in two places, now we just have a
is_img_extension() function which checks for extension
case-insensitively via grep.

in short:

* this checks the extension case-insensitively
* reduces some code duplication
* sxiv will print an error msg for invalid files instead of the script
  silently exiting

this is probably a tad bit slower, but IMO worth the trade.